+++ src/types/Theme.ts
@@ -1,0 +1,6 @@
+export interface Theme {
+    textcolors: {
+        header: string;
+        paragraph: string;
+    };
+};
